LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/popover-anchor-backdrop-transition-expected.txt


FAIL Anchored popover ::backdrop transitioning opacity with @starting-style assert_equals: expected "0" but got ""
